+++ apps/web/utils/Constants.ts
@@ -1,1 +1,7 @@
+import type { StaticImageData } from "next/image";
+
 export type Variant = "primary" | "secondary";
+
+export function resolveImageUrl(image: string | StaticImageData) {
+  return typeof image === "string" ? image : image.src;
+}
